nixfmt-rfc-style is the way
the top of modules is my favorite:
```nix
{ config, lib, pkgs, ... }:
let
  cfg = config.foo.bar.baz;
in
with lib;
{
  ...
}
```

the `let`, `in`, `with`, and opening `{` all being on the same
indentation just feels right. It separates out each thing instead into
its own set of lines rather than having something ugly like
```nix
in with lib; {
```
on a line on its own. The end of the `let` binding, the with, and the
opening of the attrset on the same line is just gross and confuses
things.